Add data for the example
Adds plutonium.csv which contains isotopic vector of plutonium
in the fuel, extracted from one simulation. The nuclides Np239
and U239 are included, since the simulation did not use a cooling
time after the final burnup step.

Adds spectra.csv which contains the neutron spectrum for each
step of the simulation (corresponding to the steps in the
plutonium file), measured in the target location for the
irm sample.