Update the code_demonstration notebook
Adds many markdown cells with extensive explanations of the
equations that are implemented in the functions. The ordering of
the code cells is changed slightly to accomodate the markdown
cells.

Minor adjustments to the docstrings in both the notebook and the
average_batch_irm module.

Burnup matrix and ratio function for boron are added to the notebook.